Pedido de Productos
Agregados Cambios a SQL: 06/05
Agregados Cambios de Ariel a la base de datos y eliminado Script 2
Modelo Hostal creado otra vez para evitar Bugs.

Pedido de Productos Completado.
Implementado el método para hacer filtro de DDL, puede ser implementado a más páginas con algo de trabajo
Botón de Limpiar ya no requiere que todos los campos sean llenados para ser presionado.

+++ Datos/DETALLE_RECEPCION.cs
@@ -16,7 +16,7 @@ namespace Datos
     {
         public short ID_DETALLE_RECEPCION { get; set; }
         public int CANTIDAD { get; set; }
-        public short ID_PRODUCTO { get; set; }
+        public long ID_PRODUCTO { get; set; }
         public short NUMERO_RECEPCION { get; set; }
     
         public virtual PRODUCTO PRODUCTO { get; set; }
